requirements.txt
blinker==1.6.2
click==8.1.3
colorama==0.4.6
colorlog==6.7.0
colormap==1.0.4
convcolors==2.2.0
easydev==0.12.1
extcolors==1.0.0
Flask==2.3.2
Flask-Cors==3.0.10
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
pexpect==4.8.0
Pillow==9.5.0
ptyprocess==0.7.0
six==1.16.0
Werkzeug==2.3.3
